Conflicts:
	framework/cc/init.lua
	framework/display.lua
	framework/functions.lua
	framework/init.lua
	lib/cocos2d-x/cocos2dx/base_nodes/CCNode.cpp
	lib/cocos2d-x/scripting/lua/cocos2dx_support/LuaCocos2d.cpp
	lib/framework_precompiled/framework_precompiled.zip
	lib/luabinding/cocos2dx/base_nodes/CCNode.tolua
	lib/luabinding/extensions/cocos-ext.tolua
	lib/proj.ios/quickcocos2dx.xcodeproj/project.pbxproj
	lib/proj.mac/quickcocos2dx.xcodeproj/project.pbxproj
	lib/proj.win32/cocos2dx_2012.vcxproj
	lib/proj.win32/cocos2dx_2012.vcxproj.filters
	player/bin/mac/quick-x-player.app/Contents/MacOS/quick-x-player
	player/bin/mac/quick-x-player.app/Contents/Resources/ConsoleWindow.nib
	player/bin/mac/quick-x-player.app/Contents/Resources/CreateNewProjectDialog.nib
	player/proj.mac/AppController.mm
	player/proj.mac/AppControllerBridge.h
	player/proj.mac/AppControllerBridge.mm
	player/welcome/scripts/game.lua
	player/welcome/scripts/scenes/WelcomeScene.lua
	samples/helloworld/proj.mac/AppController.h
	samples/iap4ios/res/framework_precompiled.zip
	samples/luajavabridge/res/framework_precompiled.zip
	template/PROJECT_TEMPLATE_01/res/framework_precompiled.zip

+++ lib/cocos2d-x/scripting/lua/luajit/build_mac.sh
@@ -1,10 +1,12 @@
 #!/usr/bin/env bash
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-SRCDIR=$DIR/LuaJit-2.0.2
+SRCDIR=$DIR/LuaJit-2.0.3
 DESTDIR=$DIR/mac
 MACOSX_DEPLOYMENT_TARGET="10.6"
 
+if [ -f "$DESTDIR"/*.a ]; then
 rm "$DESTDIR"/*.a
+fi;
 
 cd $SRCDIR
 make clean
